requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
requests==2.31.0
transformers==4.35.2
torch==2.1.1
python-dotenv==1.0.0
Pillow==10.1.0 